pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "abssctl"
dynamic = ["version"]
description = "Actual Budget Multi-Instance Sync Server Admin CLI"
readme = "README.rst"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
  {name = "Ken Robinson", email = "doccyblade@homepost.org"},
]
keywords = ["actual", "budget", "admin", "cli", "turnkey"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "cryptography>=42.0.7",
]

[project.urls]
Homepage = "https://doccyblade.github.io/abssctl/"
Discussions = "https://github.com/DocCyblade/abssctl/discussions"
Issues = "https://github.com/DocCyblade/abssctl/issues"
Source = "https://github.com/DocCyblade/abssctl"
CI = "https://github.com/DocCyblade/abssctl/actions"
Changelog = "https://github.com/DocCyblade/abssctl/blob/dev/CHANGELOG.rst"
Roadmap = "https://github.com/DocCyblade/abssctl/blob/dev/docs/roadmap.rst"             
License = "https://github.com/DocCyblade/abssctl/blob/dev/LICENSE"


[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.5.5",
  "mypy>=1.11",
  "types-PyYAML>=6.0.12",
  "sphinx>=7.4",
  "pyyaml>=6.0.3",
  "sphinx-autobuild>=2024.4",
  "sphinx-rtd-theme",
  "hatch>=1.14.2",
]

[project.scripts]
abssctl = "abssctl.cli:main"

[tool.hatch.build]
include = [
  "src/abssctl",
  "README.rst",
  "CHANGELOG.rst",
  "docs/support/actual-support-matrix.rst",
  "docs/requirements/abssctl-app-specs.txt",
  "docs/requirements/node-compat.yaml",
  "docs/requirements/node-compatibility.rst",
]

[tool.hatch.build.targets.wheel]
packages = ["src/abssctl"]
core-metadata-version = "2.3"

[tool.hatch.build.targets.sdist]
include = [
  "src/abssctl",
  "README.rst",
  "CHANGELOG.rst",
  "docs/support/actual-support-matrix.rst",
  "docs/requirements/abssctl-app-specs.txt",
  "docs/requirements/node-compat.yaml",
  "docs/requirements/node-compatibility.rst",
]
core-metadata-version = "2.3"

[tool.hatch.version]
path = "src/abssctl/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
markers = [
  "slow: marks tests that take longer to run and are excluded from quick-tests by default",
  "mutation_timeout: marks tests skipped when MUTANT_UNDER_TEST is set to keep mutation runs fast"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "D", "UP", "DTZ", "T20", "B", "Q", "ANN"]
ignore = ["D203", "D213"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
strict_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[tool.coverage.run]
branch = true
source = ["abssctl"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.sphinx]
project = "abssctl"
